back/requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
attrs==22.2.0
cachetools==5.3.1
cattrs==22.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-filter==23.2
django-graphbox==1.2.6
djangorestframework==3.14.0
facebook-sdk==3.1.0
google-auth==2.22.0
graphene==3.3
graphene-django==3.1.5
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
Markdown==3.4.4
moodlepy==0.23.10
Pillow==10.0.0
promise==2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
urllib3==1.26.16
